Renaming directories and fixing vizdoom envs
