+++ enn_zoo/enn_zoo/griddly_envs/env_descriptions/clusters_entities.yaml
