+++ enn_zoo/enn_zoo/griddly_envs/images/docs_gdy_path.png
