+++ enn_zoo/enn_zoo/vizdoom_envs/scenarios/health_gathering.wad
